# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MTC Group Staff Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: role
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new staff user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve a pending user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/suspend:
    post:
      operationId: suspendUser
      tags: [users]
      summary: Suspend a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reactivate:
    post:
      operationId: reactivateUser
      tags: [users]
      summary: Reactivate a suspended user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/force-logout:
    post:
      operationId: forceLogoutUser
      tags: [users]
      summary: Force logout all sessions for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sessions terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/{id}/offboard:
    post:
      operationId: offboardUser
      tags: [users]
      summary: Offboard a staff member (terminate sessions, archive account)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User offboarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Departments ─────────────────────────────────────────────────────────────
...
  /users/bulk-action:
    post:
      operationId: bulkUserAction
      tags: [users]
      summary: Apply an action to multiple users at once (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUserActionBody"
      responses:
        "200":
          description: Bulk action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUserActionResponse"

  # ─── Devices ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        status:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        clerkUserId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - staffId
        - email
        - fullName
        - role
        - status
        - departmentId
        - departmentName
        - clerkUserId
        - createdAt
        - updatedAt
    CreateUserBody:
      type: object
      properties:
        staffId:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
      required:
        - staffId
        - email
        - fullName
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
        departmentId:
          type: ["integer", "null"]
        status:
          type: string
...
    BulkUserActionBody:
      type: object
      properties:
        userIds:
          type: array
          items:
            type: integer
        action:
          type: string
          enum: [suspend, reactivate, force_logout]
      required:
        - userIds
        - action
    BulkUserActionResponse:
      type: object
      properties:
        message:
          type: string
        affected:
          type: integer
      required:
        - message
        - affected
...